refactor input_metadata
