+++ python/llm/src/bigdl/llm/vllm/worker/worker.py
@@ -42,7 +42,7 @@ import random
 
 from bigdl.llm.vllm.config import ModelConfig, ParallelConfig, SchedulerConfig
 from bigdl.llm.vllm.models.model_loader import get_model
-from bigdl.llm.vllm.structure.input_metadata import InputMetadata
+from bigdl.llm.vllm.model_executor.input_metadata import InputMetadata
 from bigdl.llm.vllm.structure.sampling_params import SamplingParams
 from bigdl.llm.vllm.structure.sequence import SequenceData, SamplerOutput, SequenceGroupMetadata
 from bigdl.llm.utils.common import invalidInputError
